JMRI Turnout Qualification
Turnout messages from JMRI are now qualified such that the name must
indicate that the device is on the NT bus and the address is within the
allocated turnout range (i.e. it is a physical turnout on the layout).
All other turnout devices outside these parameters will require a
set/getPanelSpecificState function at the panel level in order for the
command to have any effect (this protects against an STxxx object from
aliasing as a NTxxx object — ie. Bayshore turntable can not spoof a
turnout elsewhere on the layout).